refactor: Move resizeHandleSelectors to its own module
It's a style helper function, not a design token.

- Move resizeHandleSelectors() from tokens.ts to resizeHandle.ts
- Update imports in 4 consumer files
