+++ frontend/src/components/shell/TilingLayout.css.ts
@@ -1,5 +1,5 @@
 import { globalStyle, style } from '@vanilla-extract/css'
-import { resizeHandleSelectors } from '~/styles/tokens'
+import { resizeHandleSelectors } from '~/styles/resizeHandle'
 
 export const tilingRoot = style({
   flex: 1,
